Se agregaron las validaciones del lado del servidor que faltaban en los demás servicios web

+++ FrontEnd/ws/WSEmpresas.asmx.cs
@@ -6,6 +6,8 @@ using System.Web.Services;
 using BackEnd.Datos;
 using BackEnd.Modelo;
 using System.Web.Script.Serialization;
+using System.Text.RegularExpressions;
+using System.Security;
 namespace FrontEnd.ws
 {
     /// <summary>
@@ -24,13 +26,13 @@ namespace FrontEnd.ws
         {
             if (Session["tipo_usuario"] != null)
             {
-                if (Session["tipo_usuario"].ToString() == "Administrador")
+                if (Session["tipo_usuario"].ToString() == "Administrador" || Session["tipo_usuario"].ToString() == "Empresa")
                 {
                     JavaScriptSerializer jss = new JavaScriptSerializer();
                     return jss.Serialize(new DatosContactoDao().getAllEmpresas());
                 }
             }
-            return "";
+            throw new SecurityException();
         }
         public List<DatosContacto> getAllEmpresas()
         {
@@ -43,11 +45,22 @@ namespace FrontEnd.ws
             {
                 if (Session["tipo_usuario"].ToString() == "Administrador")
                 {
-                    JavaScriptSerializer jss = new JavaScriptSerializer();
-                    return jss.Serialize(new DatosContactoDao().delete(id));
+                    try
+                    {
+                        if (id.ToString().Length < 0)
+                        {
+                            throw new Exception("Los datos proporcionados no son válidos, verifica la información");
+                        }
+                        JavaScriptSerializer jss = new JavaScriptSerializer();
+                        return jss.Serialize(new DatosContactoDao().delete(id));
+                    }
+                    catch (System.ArgumentNullException)
+                    {
+                        throw new Exception("Los datos proporcionados no son válidos, verifica la información");
+                    }
                 }
             }
-            return "";
+            throw new SecurityException();
         }
         [WebMethod(EnableSession = true)]
         public String getOne(int id_empresa)
@@ -56,11 +69,22 @@ namespace FrontEnd.ws
             {
                 if (Session["tipo_usuario"].ToString() == "Administrador")
                 {
-                    JavaScriptSerializer jss = new JavaScriptSerializer();
+                    try
+                    {
+                        if (id_empresa.ToString().Length < 0)
+                        {
+                            throw new Exception("Los datos proporcionados no son válidos, verifica la información");
+                        }
+                        JavaScriptSerializer jss = new JavaScriptSerializer();
                     return jss.Serialize(new DatosContactoDao().getOne(id_empresa));
+                    }
+                    catch (System.ArgumentNullException)
+                    {
+                        throw new Exception("Los datos proporcionados no son válidos, verifica la información");
+                    }
                 }
             }
-            return "";
+            throw new SecurityException();
         }
         [WebMethod(EnableSession = true)]
         public String getOneEmail(string email)
@@ -69,11 +93,23 @@ namespace FrontEnd.ws
             {
                 if (Session["tipo_usuario"].ToString() == "Empresa")
                 {
-                    JavaScriptSerializer jss = new JavaScriptSerializer();
-                    return jss.Serialize(new DatosContactoDao().getOneEmail(email));
+                    try
+                    {
+                        Regex validarEmail = new Regex("^[a-zA-Z0-9._-]+@[a-zA-Z0-9.-]+\\.([a-zA-Z]{2,4})+");
+                        if (!validarEmail.IsMatch(email))
+                        {
+                            throw new Exception("Los datos proporcionados no son válidos, verifica la información");
+                        }
+                        JavaScriptSerializer jss = new JavaScriptSerializer();
+                        return jss.Serialize(new DatosContactoDao().getOneEmail(email));
+                    }
+                    catch (System.ArgumentNullException)
+                    {
+                        throw new Exception("Los datos proporcionados no son válidos, verifica la información");
+                    }
                 }
             }
-            return "";
+            throw new SecurityException();
         }
         [WebMethod(EnableSession = true)]
         public int getDatos(String email)
@@ -82,12 +118,24 @@ namespace FrontEnd.ws
             {
                 if (Session["tipo_usuario"].ToString() == "Administrador" || Session["tipo_usuario"].ToString() == "Empresa")
                 {
-                    JavaScriptSerializer jss = new JavaScriptSerializer();
-                    int resultado = new DatosContactoDao().getDatos(email);
-                    return resultado;
+                    try
+                    {
+                        Regex validarEmail = new Regex("^[a-zA-Z0-9._-]+@[a-zA-Z0-9.-]+\\.([a-zA-Z]{2,4})+");
+                        if (!validarEmail.IsMatch(email))
+                        {
+                            throw new Exception("Los datos proporcionados no son válidos, verifica la información");
+                        }
+                        JavaScriptSerializer jss = new JavaScriptSerializer();
+                        int resultado = new DatosContactoDao().getDatos(email);
+                        return resultado;
+                    }
+                    catch (System.ArgumentNullException)
+                    {
+                        throw new Exception("Los datos proporcionados no son válidos, verifica la información");
+                    }
                 }
             }
-            return 0;
+            throw new SecurityException();
         }
     }
 }
